refactor(profile): move `@top/@genres` page to `views`
