--- app/profile/[id]/@top/@genres/loading.tsx
@@ -1,5 +1,0 @@
-import { TopGenresSectionSkeleton } from '@app/profile/sections'
-
-export default function ProfileTopGenresSubLoading() {
-  return <TopGenresSectionSkeleton length={10} />
-}
