--- app/profile/[id]/@top/@genres/page.tsx
@@ -1,84 +1,0 @@
-import { redirect } from 'next/navigation'
-
-import { getUser } from '@app/api/fetchers'
-import { getRigtchTopGenres } from '@app/api/fetchers/stats/rigtch'
-import { getSpotifyTopGenres } from '@app/api/fetchers/stats/spotify'
-import type { RigtchStatsResponse } from '@app/api/types'
-import { getServerToken } from '@app/auth/utils'
-import { SeeMoreButton } from '@app/components/common/buttons'
-import {
-  STATS_MEASUREMENT,
-  STATS_PROVIDER,
-  TIME_RANGE,
-} from '@app/profile/constants'
-import {
-  StatsProvider,
-  type RigtchTimeRange,
-  type SpotifyTimeRange,
-} from '@app/profile/enums'
-import { TopGenresSection } from '@app/profile/sections'
-import type { ProfilePageProps } from '@app/profile/types'
-import { afterParamFactory } from '@app/profile/utils/factories'
-import {
-  validateStatsMeasurement,
-  validateStatsProvider,
-  validateTimeRange,
-} from '@app/profile/utils/validators'
-import { validateId } from '@app/utils/validators'
-import { isPublicUser } from '@app/profile/utils/helpers'
-
-export const runtime = 'edge'
-
-export default async function ProfileTopGenresSubPage({
-  searchParams,
-  params,
-}: ProfilePageProps) {
-  const limit = 20
-
-  const token = await getServerToken()
-  const userId = validateId(params.id)
-
-  if (!token && !isPublicUser(userId)) redirect('/')
-
-  const { createdAt } = await getUser(token ?? '', {
-    userId,
-  })
-
-  const statsProvider = validateStatsProvider(
-    searchParams[STATS_PROVIDER],
-    createdAt
-  )
-  const statsMeasurement = validateStatsMeasurement(
-    searchParams[STATS_MEASUREMENT]
-  )
-  const timeRange = validateTimeRange(
-    searchParams[TIME_RANGE],
-    statsProvider,
-    createdAt
-  )
-
-  let items: string[] | RigtchStatsResponse<string>
-
-  if (statsProvider === StatsProvider.RIGTCH) {
-    items = await getRigtchTopGenres(token ?? '', {
-      after: afterParamFactory(timeRange as RigtchTimeRange),
-      userId,
-      limit,
-      measurement: statsMeasurement,
-    })
-  } else {
-    const { genres } = await getSpotifyTopGenres(token ?? '', {
-      timeRange: timeRange as SpotifyTimeRange,
-      userId,
-      limit,
-    })
-
-    items = genres
-  }
-
-  return (
-    <TopGenresSection items={items}>
-      <SeeMoreButton href={`/profile/${userId}/top/genres`} />
-    </TopGenresSection>
-  )
-}
