+++ app/profile/[id]/views/index.ts
@@ -1,0 +1,1 @@
+export * from './top-genres-view'
